Alert requesting info on win
On win an alert will show giving the time the player completed the game in and asking for their name to add to the high scores table. After giving a name or skipping this part a button will become available to take the player back to the home screen.

A how to play page has also been added. The content comes from a text file but is not yet finalised.